# Conflicts:
#	source/components/player/comp_player_model.cpp

+++ source/components/controllers/comp_fake_animation_controller.cpp
@@ -29,7 +29,7 @@ void TCompFakeAnimation::update(float dt) {
 		pitch = 0.f;
 	}
 	else if (_animationName == "idleAction") {
-		yaw = M_2_PI * _time;
+		yaw = (float)(M_2_PI) * _time;
 	}
 	else if (_animationName == "jump") {
 		const float duration = 1.f;
